Reworked the rendering system to prep for multi-threading. The renderer now sorts sprites automatically before rendering.

+++ src/libs/core/rect.h
@@ -1,13 +1,12 @@
 #pragma once
 
-#include <inttypes.h>
 #include <stdbool.h>
 
 typedef struct rect {
-        int32_t x, y, w, h;
+        float x, y, w, h;
 } rect;
 
-static const rect rect_zero = {0, 0, 0, 0};
+static const rect rect_zero = {0.0f, 0.0f, 0.0f, 0.0f};
 
-// Returns true if the specified rect is empty (that is 0,0,0,0).
+// Returns true if the specified rect is empty (that is 0.0f,0.0f,0.0f,0.0f).
 bool rect_empty(rect*);
